\name{jdcov-package}
\alias{jdcov-package}
\alias{jdcov}
\docType{package}
\title{
\packageTitle{jdcov}
}
\description{
\packageDescription{jdcov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jdcov}
\packageIndices{jdcov}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{jdcov}

Maintainer: \packageMaintainer{jdcov}
}
\references{
Chakraborty, S. and Zhang, X. (2019). Distance Metrics for Measuring Joint Dependence with Application to Causal Inference, Journal of the American Statistical Association, DOI: 10.1080/01621459.2018.1513364.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
